refactor(iocuak): move ContainerModuleServiceImplementation to a cuaktask folder
