+++ packages/iocuak/src/container/modules/api/ContainerApi.spec.ts
@@ -15,8 +15,8 @@ jest.mock('../../../task/modules/TaskBuilder');
 jest.mock('../../../task/modules/TaskBuilderWithNoDependencies');
 jest.mock('../../../task/modules/TaskDependencyEngine');
 jest.mock('../../services/cuaktask/ContainerInstanceServiceImplementation');
+jest.mock('../../services/cuaktask/ContainerModuleServiceImplementation');
 jest.mock('../../services/domain/ContainerBindingServiceImplementation');
-jest.mock('../../services/domain/ContainerModuleServiceImplementation');
 jest.mock('../../services/domain/ContainerRequestServiceImplementation');
 jest.mock('../../services/domain/ContainerSingletonServiceImplementation');
 
@@ -32,8 +32,8 @@ import { TaskBuilder } from '../../../task/modules/TaskBuilder';
 import { TaskBuilderWithNoDependencies } from '../../../task/modules/TaskBuilderWithNoDependencies';
 import { TaskDependencyEngine } from '../../../task/modules/TaskDependencyEngine';
 import { ContainerInstanceServiceImplementation } from '../../services/cuaktask/ContainerInstanceServiceImplementation';
+import { ContainerModuleServiceImplementation } from '../../services/cuaktask/ContainerModuleServiceImplementation';
 import { ContainerBindingServiceImplementation } from '../../services/domain/ContainerBindingServiceImplementation';
-import { ContainerModuleServiceImplementation } from '../../services/domain/ContainerModuleServiceImplementation';
 import { ContainerRequestServiceImplementation } from '../../services/domain/ContainerRequestServiceImplementation';
 import { ContainerSingletonServiceImplementation } from '../../services/domain/ContainerSingletonServiceImplementation';
 import { ContainerApi } from './ContainerApi';
